==> 0_preprocessing.ipynb <==
Before we use the dataset, we are going to filter out some songs to both get the data to a more manageable size and exclude songs we don't want to include in our search engine (ex. non-english songs)

import pandas as pd
import regex as re

# mount google drive
from google.colab import drive
drive.mount('/content/gdrive')

# note: place shortcut to shared project folder in google drive root directory
%cd /content/gdrive/MyDrive/ml2_project

Reading in the whole dataset takes ~9 min on my computer

#read first 10 million rows (all of the dataset)
whole_df = pd.read_csv("data/ds2.csv", nrows = 10_000_000)

len(whole_df)

gt_million = sum(whole_df['views'] > 1_000_000)
gt_hundred_thousand = sum(whole_df['views'] > 100_000)
gt_ten_thousand = sum(whole_df['views'] > 10_000)
gt_million, gt_hundred_thousand, gt_ten_thousand

pcts = [n / len(whole_df) for n in (gt_million, gt_hundred_thousand, gt_ten_thousand)]
pcts = [100 * p for p in pcts]
pcts

It makes sense to filter songs based on the number of views, since songs with few views are more likey to pollute search results than actually be searched for.

Filtering for lyrics with more than 1,000,000, 100,000, and 10,000 leaves 1,396, 23,698, and 185,791 songs resp. which is 0.02%, 0.4 % and 3.14% of the original dataset size

We are going to use only songs with > 10,000 views

df = whole_df[whole_df['views'] > 10_000].copy().reset_index()

Some of the songs in this table don't have lyrics, so we will drop those

df = df[~df['lyrics'].isna()].copy()

sum(df['lyrics'].isna())

The lyrics often contain annotations that we don't want to analyze, so we can remove them using regex:

#limit size of brackets to 100 so the regex doesn't stall out on unclosed brackets
df['lyrics'] = df['lyrics'].replace("\[.{,100}]", "", regex = True)

lyrics = df['lyrics']
lyrics

We also want to remove foreign language songs, the best we can do on this front is remove any song with a non-english character.

non_eng_pattern = '[^\x00-\xFF]'

def to_eng(_lyrics, find_char, replace_char):
    _lyrics = _lyrics.replace(find_char, replace_char, regex = True)
    print(f"Changing {find_char} to {replace_char}.")
    new_sum = sum(_lyrics.str.contains(non_eng_pattern))
    print(f"reduced non-eng songs to {new_sum} \n")
    return(_lyrics)

len(df)

sum(lyrics.str.contains(non_eng_pattern))

sum(lyrics.str.contains(non_eng_pattern)) / len(df)

About 42% of our songs contain non-latin characters, but it turns out a lot of that is due to the curious use of lookalikes for what should be normal ASCII characters. We can again use regex to replace these lookalikes.

replacements = [("[—–‒−]", "-"), ("[‘’′ʼ]", "'"),("…","..."), ('[“”″]', '"'),("[\u2005\u205f\xa0\u200a]", " "),
                ("[\u200b\u2060\u200e\ufeff]", ""), ("е", "e"), ("‚",","), ("Ι", "I")]
for rep in replacements:
    lyrics = to_eng(lyrics, rep[0], rep[1])
df['lyrics'] = lyrics

sum(lyrics.str.contains(non_eng_pattern)) / len(df)

When we replace the lookalikes, the percentage of "non-english" songs goes down from 42% to 14%. Particularly, there seems to be a lot of obscure utf whitespace characters used.

Upon further inspection we can confirm our method sorts between English and Non-English songs about as well as we can expect to.

eng_df = df[~df['lyrics'].str.contains(non_eng_pattern)]
eng_df

non_eng_df = df[df['lyrics'].str.contains(non_eng_pattern)]
non_eng_df

Still, a non-significant portion of the songs we filtered as not English appear to be English. We can look at why this is the case:

def show_non_eng_lyric(lyric_str, half_width):
    non_eng = re.search(non_eng_pattern, lyric_str)
    left_bound = non_eng.span()[0] - half_width
    right_bound = non_eng.span()[1] + half_width
    print(lyric_str[max(left_bound, 0):min(right_bound, len(lyric_str)- 1)])

for i in range(len(non_eng_df)):
    print(f"\n{i+1}")
    print(f"{non_eng_df['title'].iloc[i]} by {non_eng_df['artist'].iloc[i]}")
    show_non_eng_lyric(non_eng_df['lyrics'].iloc[i], 20)

for i in range(len(non_eng_df)):
    print(re.search(non_eng_pattern, non_eng_df['lyrics'].iloc[i]))

A lot of the lyrics we wrongly filtered as non-english use accents that aren't on our character "whitelist".

In the future we can go back and get more particular with out methods to filter non-english songs, but for now I consider this a reasonable amount. (By inspection about 15% of filtered songs are English)

# Write out clean datasets

We are going to write 2 datasets, one with all the songs with greater than 10,000 views that we plan to use in our final implementation of the project, and one with songs greater than 1,000,000 views that we can play around with a bit more easily.

mini_df = df[df['views'] > 1_000_000]

mini_df.to_csv("data/small_dataset.csv")

df.to_csv("data/clean_dataset.csv")
